Move test that cannot be used in unittest class to own file
